market-analysis-apllication/requirements.txt
Flask
yfinance
pandas
serpapi
google-generativeai
